Apply more patches from Sven Eckelmann:
 * Respect (un)signess of parameter in printf format strings
 * Don't convert uint64_t to float when only comparing against 0
 * Don't convert new_hna_len from int16_t to int and back to int16_t



git-svn-id: http://downloads.open-mesh.org/svn/batman/trunk/batman@1168 45894c77-fb22-0410-b583-ff6e7d5dbf6c

+++ batman.c
@@ -255,7 +255,7 @@ static int hna_buff_delete(struct hna_element *buf, int *buf_len, struct hna_ele
  * a situation where no route is present.
  */
 void update_hna(struct orig_node *orig_node, unsigned char *new_hna,
-				int new_hna_len, struct neigh_node *old_router)
+				int16_t new_hna_len, struct neigh_node *old_router)
 {
 	unsigned char *old_hna;
 	int old_hna_len;
